requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0
gTTS==2.5.0
python-multipart==0.0.9
starlette==0.36.3
aiofiles==23.2.1